refactor: extract postfetch utilities to standalone module and simplify `default_qs_hook` logic by delegating postfetch optimizations
